Re-add PersonaGenerator abstract class and fix: stop-tokens in output
Renamed TGIPersonaGenerator to ExpertGenerator
