+++ mallm/coordinator.py
@@ -22,7 +22,7 @@ from mallm.discourse_policy.DiscourceRelay import DiscourseRelay
 from mallm.discourse_policy.DiscourceReport import DiscourseReport
 from mallm.discourse_policy.DiscoursePolicy import DiscoursePolicy
 from mallm.models.Chat import Chat
-from mallm.models.personas.PersonaGenerator import PersonaGenerator
+from mallm.models.personas.ExpertGenerator import ExpertGenerator
 from mallm.prompts.coordinator_prompts import generate_chat_prompt_extract_result
 from mallm.utils.types import Agreement, Memory
 
@@ -48,7 +48,7 @@ class Coordinator:
         self,
         model: Chat,
         client: httpx.Client,
-        agent_generator: Optional[PersonaGenerator] = None,
+        agent_generator: Optional[ExpertGenerator] = None,
         use_moderator: bool = False,
         memory_bucket_dir: str = "./mallm/utils/memory_bucket/",
     ):
